Corrigindo metodos e refatorando a classe

+++ src/main/java/com/redelog/api/repository/EntregadorRepository.java
@@ -3,5 +3,5 @@ package com.redelog.api.repository;
 import com.redelog.api.model.entities.Entregador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntregadorRepository extends JpaRepository<Entregador, Integer> {
+public interface EntregadorRepository extends JpaRepository<Entregador, Long> {
 }
